// src/types/models/FavoriteItemsResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ProductResponse } from './ProductResponse';
export type FavoriteItemsResponse = {
    id: number;
    favorite_id: number;
    product: ProductResponse;
};